+++ lib/types/simulation.ts
@@ -1,0 +1,10 @@
+export type PulseSpec = {
+  v1: number
+  v2: number
+  td: number
+  tr: number
+  tf: number
+  ton: number
+  period: number
+  ncycles: number
+}
